requirements.txt
bcrypt==3.2.0
cffi==1.14.6
click==8.0.1
dnspython==1.16.0
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-PyMongo==2.3.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pycparser==2.20
pymongo==3.12.0
python-decouple==3.4
six==1.16.0
Werkzeug==2.0.1
